refactor: restructure and renaming
