--- ee/packages/amplication-git-push-webhook-service/src/contracts/queue.interface.ts
@@ -1,5 +1,0 @@
-import { CreateRepositoryPushRequest } from '../entities/dto/CreateRepositoryPushRequest';
-
-export interface QueueInterface {
-  createPushRequest(createRepositoryPushRequest: CreateRepositoryPushRequest);
-}
