--- ee/packages/amplication-git-push-webhook-service/src/entities/enums/provider.ts
@@ -1,3 +1,0 @@
-export enum EnumProvider {
-  Github = 'Github',
-}
