+++ ee/packages/amplication-git-push-webhook-service/src/git-organization/git-organization.repository.ts
@@ -1,7 +1,7 @@
-import { GitOrganization, PrismaService } from '@amplication/prisma-db';
+import { GitOrganization, PrismaService } from '@amplication/prisma-clients/amplication-prisma-db';
 import { Injectable } from '@nestjs/common';
-import { GitOrganizationInterface } from '../contracts/gitOrganization.interface';
-import { EnumProvider } from '../entities/enums/provider';
+import { EnumProvider, GitOrganizationInterface } from './git-organization.types';
+
 
 @Injectable()
 export class GitOrganizationRepository implements GitOrganizationInterface {
