--- ee/packages/amplication-git-push-webhook-service/src/modules/root-winston.module.ts
@@ -1,9 +1,0 @@
-import { ConfigModule } from '@nestjs/config';
-import { WinstonModule } from 'nest-winston';
-import { WinstonConfigService } from '../services/winstonConfig.service';
-
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export const RootWinstonModule = WinstonModule.forRootAsync({
-  imports: [ConfigModule],
-  useClass: WinstonConfigService,
-});
