+++ ee/packages/amplication-git-push-webhook-service/src/queue/dto/repository-push-create-event.ts
@@ -1,4 +1,5 @@
-import { EnumProvider } from '../enums/provider';
+import { EnumProvider } from "../../git-organization/git-organization.types";
+
 
 export class RepositoryPushCreateEvent {
   constructor(
